+++ tests/testthat/test-mlp-binary.R
@@ -68,7 +68,7 @@ test_that('predictions', {
   x_df <- as.data.frame(scale(x_df))
 
   set.seed(1)
-  fit_df <- brulee_mlp(x_df, y, epochs = 10L, batch_size = length(y))
+  fit_df <- brulee_mlp(x_df, y, epochs = 10L, batch_size = 32, optimizer = "SGD")
 
   complete_pred <- predict(fit_df, head(x_df))
   expect_true(tibble::is_tibble(complete_pred))
@@ -99,7 +99,6 @@ test_that("mlp binary learns something", {
   y <- as.factor(x > 0)
 
   model <- brulee_mlp(x, y,
-                      batch_size = 50,
                       epochs = 100L,
                       activation = "relu",
                       hidden_units = 5L,