+++ NEWS.md
@@ -1,5 +1,7 @@
 # brulee (development version)
 
+* An `optimizer` was added to [brulee_mlp()], with a new default being LBFGS instead of stochastic gradient descent. 
+
 # brulee 0.1.0
 
 * Modeling functions gained a `mixture` argument for the proportion of L1 penalty that is used. (#50)